// apps/web/generated.ts
/**
 * Generated by orval v6.10.3 🍺
 * Do not edit manually.
 * Fastify Template
 * Swagger Spec for Fastify web API
 * OpenAPI spec version: 0.1.0
 */
import { appInstance } from "./utils/custom-axios";
export type PostApiUsersRegisterBody = {
  name: string;
  email: string;
  password: string;
  password2: string;
};

export type GetApiUsersMe200 = {
  id: string;
  created_at: string;
  email?: string;
  phone?: string;
  confirmed_at?: string;
  email_confirmed_at?: string;
  phone_confirmed_at?: string;
};

/**
 * API Information
 */
export type GetApi200 = {
  version: string;
};

/**
 * Health check
 */
export type GetHeartbeat200 = {
  health: string;
};

// eslint-disable-next-line
type SecondParameter<T extends (...args: any) => any> = T extends (
  config: any,
  args: infer P
) => any
  ? P
  : never;

export const getHeartbeat = (options?: SecondParameter<typeof appInstance>) => {
  return appInstance<GetHeartbeat200>(
    { url: `/heartbeat`, method: "get" },
    options
  );
};

export const getApi = (options?: SecondParameter<typeof appInstance>) => {
  return appInstance<GetApi200>({ url: `/api/`, method: "get" }, options);
};

export const getApiUsersMe = (
  options?: SecondParameter<typeof appInstance>
) => {
  return appInstance<GetApiUsersMe200>(
    { url: `/api/users/me`, method: "get" },
    options
  );
};

export const postApiUsersRegister = (
  postApiUsersRegisterBody: PostApiUsersRegisterBody,
  options?: SecondParameter<typeof appInstance>
) => {
  return appInstance<void>(
    {
      url: `/api/users/register`,
      method: "post",
      headers: { "Content-Type": "application/json" },
      data: postApiUsersRegisterBody,
    },
    options
  );
};

export type GetHeartbeatResult = NonNullable<
  Awaited<ReturnType<typeof getHeartbeat>>
>;
export type GetApiResult = NonNullable<Awaited<ReturnType<typeof getApi>>>;
export type GetApiUsersMeResult = NonNullable<
  Awaited<ReturnType<typeof getApiUsersMe>>
>;
export type PostApiUsersRegisterResult = NonNullable<
  Awaited<ReturnType<typeof postApiUsersRegister>>
>;